Test environment for xp_fifo.
There are two issues in previous xp_fifo design that need attention.

First, when pre-fetch FIFO is currently full but unloading a data, the input data(c_data) of xp_fifo at this clock cycle can bypass the memory and be sent into pre-fetch FIFO directly, given that the unloaded data will leave a spot for the incoming one. This issue has been fixed by including this condition when deciding whether to bypass the memory or not when a new transaction on consumer's side happens.

Second, when a data is being read out from memory, it needs three clock cycle latency to be written into pre-fetch FIFO.  During this period, the input c_data will be written into memory rather than pre-fetch FIFO to maintain the order of data sequence even if pre-fetch FIFO is not full. Therefore, the c_usage could be larger than 0 but in fact no data is ready at pre-fetch FIFO. So p_srdy = 0 when c_usage>0  is a possible case. To solve this problem, two write pointers will be needed, one for memory data currently being read and one for input c_data. Then, reserving an entry in pre-fetch FIFO for data from memory by assigning the c_data write pointer to next address and write input data starting from there.This problem is not being resolved at this moment due to the complexity and time involved.